Fix rulenames which includes an index
The rulename has a regex pattern `[/^[a-zA-Z0-9_]+-[a-zA-Z0-9_]+(-\d+)?$/]`
which allows an index at the end of the rulename (with a delimiter).
This is split later with `$data = split($rulename, '-')` but the content
of the index part is never used as a fragment. Therefor the fragment was
defined twice.
This fix will add the index to the fragment name.