Provide a mechanism to flush un-managed rules
This patchset adds a new parameter to the main class to activate a
mechanism that will invoke `systemctl reload nftables` during the
Puppet run if manual changes to the in-memory ruleset are detected.

To accomplish this, the systemd unit in charge of nftables is
configured to write a hash of the in-memory ruleset right after
starting/reloading. During the Puppet run, the hash of the current
rule set is compared to the one previously stored. If the hash differ
then `systemctl reload nftables` is executed to flush manual changes.

Fixes #113