Make sure that the list of interfaces is mocked
If 'networking' is not removed from the facts, the values declared in
`default_module_facts.yml` win (the fact set contains two keys:
`:networking` and `'networking'`)

This test has always been wrong as the interfaces were never been
mocked and the asserts were based on the contents on the default
module facts, not the test context itself.

The bug is obvious in the Puppet6/CentOS9 combination as there's no
`networking` fact in FacterDB, hence only in that case the fact is
overridden and the test does not pass (as it shouldn't).

This fix makes sure that the fact set in the test context is correct
and also that the test declares the actually expected resources.